fix(user): 修改http请求的proto文件，加入了userId从上下文中获取的逻辑

+++ app/auth/biz/service/verify_token_by_rpc.go
@@ -74,7 +74,7 @@ func (s *VerifyTokenByRPCService) Run(req *auth.VerifyTokenReq) (resp *auth.Veri
 
 	return &auth.VerifyResp{
 		Res:    true,
-		UserId: int32(claims.UserId),
+		UserId: claims.UserId,
 		Role:   claims.Role,
 	}, nil
 }
